// reactjs/reactjs-10/src/models/types.ts
export enum ActionTypes {
  SIGNIN = 'APP/SIGNIN',
  LOGOUT = 'APP/LOGOUT',
  REGISTRATION = 'APP/REGISTRATION',
  ONCHAGEAUTHFIELD = 'AUTH/ONCHAGEAUTHFIELD',
  ONCHAGEREGISTRATIONFIELD = 'AUTH/ONCHAGEREGISTRATIONFIELD',
  DISABLEDLOGIN = 'APP/DISABLEDLOGIN',
  ENABLEDLOGIN = 'APP/ENABLEDLOGIN',
  DISABLEDREGISTRATION = 'APP/DISABLEDREGISTRATION',
  ENABLEDREGISTRATION = 'APP/ENABLEDREGISTRATION',
  NOTIFICATION = 'APP/NOTIFICATION',
  TODOCHANGEFIELD = 'TODO/TODOCHANGEFIELD',
  TODOADD = 'TODO/TODOADD',
  TODOMARKED = 'TODO/TODOMARKED',
  TODODELETE = 'TODO/TODODELETE',
}